Put all test constants in a factory
