+++ src/test/kotlin/helper/LogFactory.kt
@@ -9,6 +9,7 @@ import logic.entities.LogItem
 import java.util.*
 
 object LogFactory {
+    private val NOW = Clock.System.now().toLocalDateTime(TimeZone.Companion.UTC)
 
     fun createLogItem(
         id: UUID = UUID.randomUUID(),
@@ -21,7 +22,6 @@ object LogFactory {
         entityId = entityId,
         date = date
     )
-    private val NOW = Clock.System.now().toLocalDateTime(TimeZone.Companion.UTC)
 
     val LOG_1 = LogItem(
         id = UUID.fromString("aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa"),
@@ -38,7 +38,8 @@ object LogFactory {
     )
 
     val LOGS_FOR_PROJECT_1 = listOf(LOG_1, LOG_2)
-    fun taskLogs ()= listOf<LogItem>(
+
+    fun taskLogs() = listOf<LogItem>(
         LogItem(
             id = UUID.fromString("11111111-1111-1111-1111-111111111111"),
             message = "This task name is updated by Hend at 12:30 29-4-2025",
@@ -52,11 +53,4 @@ object LogFactory {
             entityId = UUID.fromString("123e4567-e89b-12d3-a456-426614174000"),
         )
     )
-
-    val validId = UUID.fromString("123e4567-e89b-12d3-a456-426614174000")
-    val invalidId = "123e4567-e89b-12d3-a456-426614174%@0"
-
-    const val TASK_NOT_FOUND = "Task not found. Please make sure the task ID is correct."
-    const val INVALID_ID_FORMAT = "This Id is invalid format , Please ensure you enter correct format of id"
-
 }
