--- src/test/kotlin/helper/LoginTestFactory.kt
@@ -1,9 +1,0 @@
-package helper
-
-
-
-
-
-const val SUCCESS_MESSAGE_FOR_LOGIN="Success Login......"
-const val WRONG_USER_NAME="The user name you entered is wrong , Please enter correct user name"
-const val WRONG_PASSWORD="The password you entered is wrong , Please enter correct password"
