+++ src/test/kotlin/logic/usecases/logs/ViewTaskLogsUseCaseTest.kt
@@ -1,8 +1,8 @@
 package logic.usecases.logs
 
 import com.google.common.truth.Truth
+import helper.ConstantsFactory.validId
 import helper.LogFactory.taskLogs
-import helper.LogFactory.validId
 import io.mockk.coEvery
 import io.mockk.every
 import io.mockk.mockk
