+++ src/test/kotlin/logic/usecases/state/GetProjectStatesUseCaseTest.kt
@@ -8,8 +8,8 @@ import logic.exception.PlanMateException.ValidationException.InvalidProjectIDExc
 import logic.exception.PlanMateException.NotFoundException.ProjectNotFoundException
 import logic.repository.ProjectRepository
 import helper.ProjectFactory.dummyProject
-import helper.ProjectFactory.EXISTING_PROJECT_ID
-import helper.ProjectFactory.INVALID_PROJECT_ID
+import helper.ConstantsFactory.EXISTING_PROJECT_ID
+import helper.ConstantsFactory.INVALID_PROJECT_ID
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertThrows
