+++ src/test/kotlin/logic/usecases/utils/ValidateInputUseCaseTest.kt
@@ -1,7 +1,7 @@
 package logic.usecases.utils
 
 import com.google.common.truth.Truth
-import helper.ValidateInputUseCaseTestFactory
+import helper.ConstantsFactory
 import org.junit.jupiter.api.Test
 
 class ValidateInputUseCaseTest {
@@ -12,7 +12,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return true when take a valid name`() {
         //Given & When
-        val result = validateInputUseCase.isValidName(ValidateInputUseCaseTestFactory.VALID_NAME)
+        val result = validateInputUseCase.isValidName(ConstantsFactory.VALID_NAME)
 
         //Then
         Truth.assertThat(result).isTrue()
@@ -21,7 +21,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return false when take a blank name`() {
         //Given & When
-        val result = validateInputUseCase.isValidName(ValidateInputUseCaseTestFactory.BLANK_NAME)
+        val result = validateInputUseCase.isValidName(ConstantsFactory.BLANK_NAME)
 
         //Then
         Truth.assertThat(result).isFalse()
@@ -30,7 +30,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return false when take a short name(less than 3) characters`() {
         //Given & When
-        val result = validateInputUseCase.isValidName(ValidateInputUseCaseTestFactory.SHORT_NAME)
+        val result = validateInputUseCase.isValidName(ConstantsFactory.SHORT_NAME)
 
         //Then
         Truth.assertThat(result).isFalse()
@@ -39,7 +39,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return false when take a name longer than 100 characters`() {
         //Given & When
-        val result = validateInputUseCase.isValidName(ValidateInputUseCaseTestFactory.LONG_NAME)
+        val result = validateInputUseCase.isValidName(ConstantsFactory.LONG_NAME)
 
         //Then
         Truth.assertThat(result).isFalse()
@@ -48,7 +48,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return false when take a name with invalid characters`() {
         //Given & When
-        val result = validateInputUseCase.isValidName(ValidateInputUseCaseTestFactory.NAME_WITH_INVALID_CHARACTERS)
+        val result = validateInputUseCase.isValidName(ConstantsFactory.NAME_WITH_INVALID_CHARACTERS)
 
         //Then
         Truth.assertThat(result).isFalse()
@@ -57,7 +57,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return true when take a name with underscores, spaces`() {
         //Given & When
-        val result = validateInputUseCase.isValidName(ValidateInputUseCaseTestFactory.NAME_WITH_UNDERSCORES_AND_SPACES)
+        val result = validateInputUseCase.isValidName(ConstantsFactory.NAME_WITH_UNDERSCORES_AND_SPACES)
 
         //Then
         Truth.assertThat(result).isTrue()
@@ -70,7 +70,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return true when take a valid UUID`() {
         //Given & When
-        val result = validateInputUseCase.isValidUUID(ValidateInputUseCaseTestFactory.VALID_UUID)
+        val result = validateInputUseCase.isValidUUID(ConstantsFactory.VALID_UUID)
 
         //Then
         Truth.assertThat(result).isTrue()
@@ -79,7 +79,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return false when take an invalid UUID`() {
         //Given & When
-        val result = validateInputUseCase.isValidUUID(ValidateInputUseCaseTestFactory.INVALID_UUID_STRING_FORMAT)
+        val result = validateInputUseCase.isValidUUID(ConstantsFactory.INVALID_UUID_STRING_FORMAT)
 
         //Then
         Truth.assertThat(result).isFalse()
@@ -88,7 +88,7 @@ class ValidateInputUseCaseTest {
     @Test
     fun `should return false when take an empty string as UUID`() {
         //Given & When
-        val result = validateInputUseCase.isValidUUID(ValidateInputUseCaseTestFactory.EMPTY_STRING)
+        val result = validateInputUseCase.isValidUUID(ConstantsFactory.EMPTY_STRING)
 
         //Then
         Truth.assertThat(result).isFalse()
