+++ src/test/kotlin/ui/login/LoginUIControllerTest.kt
@@ -1,10 +1,10 @@
 package ui.login
 
 import com.google.common.truth.Truth
-import helper.SUCCESS_MESSAGE_FOR_LOGIN
+import helper.ConstantsFactory.SUCCESS_MESSAGE_FOR_LOGIN
 import helper.UserFactory.validUserData
-import helper.WRONG_PASSWORD
-import helper.WRONG_USER_NAME
+import helper.ConstantsFactory.WRONG_PASSWORD
+import helper.ConstantsFactory.WRONG_USER_NAME
 import io.mockk.CapturingSlot
 import io.mockk.Runs
 import io.mockk.coEvery
